Call GC after each training step.

+++ src/main/scala/lambdanet/utils/FileLogger.scala
@@ -43,7 +43,7 @@ object EventLogger {
 
   case class MapValue(v: Vector[(String, Double)]) extends EventValue
 
-  case class Event(name: String, iteration: Int, value: EventValue)
+  case class Event(name: String, iteration: Double, value: EventValue)
 
   case class PlotConfig(options: String*)
 }
@@ -66,7 +66,7 @@ class EventLogger(
     ???
   }
 
-  def logScalar(name: String, iteration: Int, value: Double): Unit = {
+  def logScalar(name: String, iteration: Double, value: Double): Unit = {
     log(Event(name, iteration, DoubleValue(value)))
   }
